Match uniform precision across the ring shader stages
'radius' was highp in the vertex shader (the GLSL default) and
mediump in the fragment shader; GLSL ES 300 rejects the mismatch at
link time. The TileRenderer constructor threw on every mount since
the ring pass landed, which is what has been blanking the viewer in
roll mode. Caught the moment webview errors reached the terminal.

+++ app/src/lib/renderer.ts
@@ -52,7 +52,10 @@ void main() {
 }`;
 
 const RING_FRAG = `#version 300 es
-precision mediump float;
+// highp to match the vertex stage: 'radius' is declared in both shaders and
+// GLSL ES 300 requires identical precision, or the program fails to link
+// (vertex-stage floats default to highp). WebGL2 guarantees fragment highp.
+precision highp float;
 in vec2 vCorner;
 out vec4 color;
 uniform float radius;
